+++ .env_example
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-export FLASK_APP=metabulo/wsgi
+export FLASK_APP=viime/wsgi
 export FLASK_ENV=development
-export SQLALCHEMY_DATABASE_URI=sqlite:///${HOME}/metabulo/db.sqlite3
+export SQLALCHEMY_DATABASE_URI=sqlite:///${HOME}/viime/db.sqlite3
 export SECRET_KEY=deadbeef
-export UPLOAD_FOLDER=${HOME}/metabulo/files
+export UPLOAD_FOLDER=${HOME}/viime/files
 export MAX_FILE_UPLOAD_SIZE=5242880
 export OPENCPU_API_ROOT=http://localhost:8004/ocpu/library
