+++ .env_pwd
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-export FLASK_APP=metabulo/wsgi
+export FLASK_APP=viime/wsgi
 export FLASK_ENV=development
-export SQLALCHEMY_DATABASE_URI=sqlite:///${PWD}/metabulo_sqlite/db.sqlite3
+export SQLALCHEMY_DATABASE_URI=sqlite:///${PWD}/viime_sqlite/db.sqlite3
 export SECRET_KEY=deadbeef
-export UPLOAD_FOLDER=${PWD}/metabulo_sqlite/files
+export UPLOAD_FOLDER=${PWD}/viime_sqlite/files
 export MAX_FILE_UPLOAD_SIZE=5242880
 export OPENCPU_API_ROOT=http://localhost:8004/ocpu/library
