+++ .gitignore
@@ -73,6 +73,7 @@ instance/
 .scrapy
 
 # Sqlite stuff:
+viime_sqlite/
 metabulo_sqlite/
 
 # Sphinx documentation
