+++ Pipfile
@@ -6,7 +6,7 @@ verify_ssl = true
 [dev-packages]
 
 [packages]
-metabulo = {editable = true,path = "."}
+viime = {editable = true,path = "."}
 
 [requires]
 python_version = "3.7"
