+++ README.md
@@ -1,11 +1,11 @@
-Metabulo
+VIIME
 ========
 
 Getting started
 ---------------
 
 This project includes a `Pipfile` to help set up a virtual environment for
-Metabulo. To set up the virtual environment
+VIIME. To set up the virtual environment
 
 ```sh
 pipenv install
@@ -17,15 +17,15 @@ and to enter it
 pipenv shell
 ```
 
-Metabulo is configured from a `.env` file present in the current directory where
+VIIME is configured from a `.env` file present in the current directory where
 it is executed.  See the included [.env_example](./.env_example) for an example
 (or try [.env_pwd](./env_pwd), which saves DB and uploaded files in the checked
 out directory rather than your home directory).  Once the environment is in
 place, you will need to initialize the tables by running
 
 ```sh
-mkdir metabulo_sqlite
-metabulo-create-tables
+mkdir viime_sqlite
+viime-create-tables
 ```
 
 This will create a data directory according to the `SQLALCHEMY_DATABASE_URI` and
@@ -54,13 +54,13 @@ for the processing backend.  To run it locally, build the docker container
 
 ```sh
 cd devops
-docker build -t metabulo .
+docker build -t viime .
 ```
 
 and start the instance
 
 ```sh
-docker run -it -p 8004:8004 metabulo
+docker run -it -p 8004:8004 viime
 ```
 
 You may find that changes to the position or size of the window in which the
@@ -69,17 +69,17 @@ actually intended behavior and can be avoided by running the image *without*
 allocating a pseudo-terminal and keeping `stdin` open:
 
 ```sh
-docker run --name metabulo -p 8004:8004 metabulo
+docker run --name viime -p 8004:8004 viime
 ```
 
-To stop this container, use a command like `docker stop metabulo` from another
+To stop this container, use a command like `docker stop viime` from another
 terminal.
 
 This installation includes a custom R package called
-[metabulo](devops/metabulo), which contains all functions that are exposed to
+[viime](devops/viime), which contains all functions that are exposed to
 the API server for CSV file processing.  Functions contained in this package
 should accept a path to a CSV file as an argument and return a data frame.  The
-(work-in-progress) backend code at [metabulo/opencpu.py](metabulo/opencpu.py)
+(work-in-progress) backend code at [viime/opencpu.py](viime/opencpu.py)
 handles the communication between pandas and R data frames.  To add additional
-methods exposed to the API server, add the function to the metabulo package and
+methods exposed to the API server, add the function to the viime package and
 rebuild the docker image.
