+++ devops/Dockerfile
@@ -1,6 +1,6 @@
-FROM metabulo/opencpu-base
+FROM viime/opencpu-base
 
-ADD metabulo /metabulo
-RUN R -e 'library(devtools); library(roxygen2); setwd("/metabulo"); document(); setwd("/"); install("metabulo")'
+ADD viime /viime
+RUN R -e 'library(devtools); library(roxygen2); setwd("/viime"); document(); setwd("/"); install("viime")'
 
 CMD service cron start && /usr/lib/rstudio-server/bin/rserver && apachectl -DFOREGROUND
