--- devops/Vagrantfile
@@ -1,26 +1,0 @@
-# -*- mode: ruby -*-
-# vi: set ft=ruby :
-
-Vagrant.configure("2") do |config|
-
-  config.vm.box = "ubuntu/bionic64"
-
-  config.vm.network "private_network", ip: "192.168.33.10"
-
-  config.vm.provider :virtualbox do |vb|
-    vb.customize ["modifyvm", :id, "--memory", 4096]
-    vb.customize ["modifyvm", :id, "--cpus", 2]
-  end
-
-  config.vm.provision "ansible" do |ansible|
-    ansible.galaxy_role_file = "ansible/requirements.yml"
-    ansible.playbook = "ansible/vagrant.yml"
-    ansible.extra_vars = {
-      metabulo_sdist_path: '../../dist/metabulo-0.1.0.tar.gz',
-      flask_secret: 'thisisasecretbetweenyouandme',
-      aws_access_key_id: 'aws_access_key_id',
-      aws_secret_access_key: 'aws_secret_access_key'
-    }
-  end
-
-end
