+++ devops/ansible/data/nginx/sites-available/viime.org.conf
@@ -2,18 +2,18 @@ server {
 	  listen 443 ssl;
 
 	  server_name viime.org;
-	  root /home/gunicorn/venvs/metabulo/share/metabulo/static;
+	  root /home/gunicorn/venvs/viime/share/viime/static;
 
 	  # SSL
 	  ssl_certificate /etc/letsencrypt/live/viime.org/fullchain.pem;
 	  ssl_certificate_key /etc/letsencrypt/live/viime.org/privkey.pem;
 
     location / {
-        try_files $uri $uri/ @metabulo;
+        try_files $uri $uri/ @viime;
     }
 
 	  # reverse proxy
-	  location @metabulo {
+	  location @viime {
 		    proxy_pass http://localhost:8000;
 		    include nginxconfig.io/proxy.conf;
 	  }
