--- devops/ansible/data/vagrant-nginx/nginx.conf
@@ -1,35 +1,0 @@
-# Generated by nginxconfig.io
-# https://nginxconfig.io/?0.redirect=false&0.https=false&0.php=false&0.proxy&0.fallback_html
-
-user nginx;
-pid /run/nginx.pid;
-worker_processes auto;
-worker_rlimit_nofile 65535;
-
-events {
-	multi_accept on;
-	worker_connections 65535;
-}
-
-http {
-	charset utf-8;
-	sendfile on;
-	tcp_nopush on;
-	tcp_nodelay on;
-	server_tokens off;
-	log_not_found off;
-	types_hash_max_size 2048;
-	client_max_body_size 16M;
-
-	# MIME
-	include mime.types;
-	default_type application/octet-stream;
-
-	# logging
-	access_log /var/log/nginx/access.log;
-	error_log /var/log/nginx/error.log warn;
-
-	# load configs
-	include /etc/nginx/conf.d/*.conf;
-	include /etc/nginx/sites-enabled/*;
-}
