--- devops/ansible/data/vagrant-nginx/nginxconfig.io/proxy.conf
@@ -1,11 +1,0 @@
-proxy_http_version	1.1;
-proxy_cache_bypass	$http_upgrade;
-
-proxy_set_header Upgrade			$http_upgrade;
-proxy_set_header Connection 		"upgrade";
-proxy_set_header Host				$host;
-proxy_set_header X-Real-IP			$remote_addr;
-proxy_set_header X-Forwarded-For	$proxy_add_x_forwarded_for;
-proxy_set_header X-Forwarded-Proto	$scheme;
-proxy_set_header X-Forwarded-Host	$host;
-proxy_set_header X-Forwarded-Port	$server_port;
