--- devops/ansible/data/vagrant-nginx/sites-available/example.com.conf
@@ -1,16 +1,0 @@
-server {
-  listen 80;
-
-  root /home/gunicorn/venvs/metabulo/share/metabulo/static; 
-
-  location / {
-      try_files $uri $uri/ @metabulo;
-  }
-
-	location @metabulo {
-	    proxy_pass http://localhost:8000;
-		  include nginxconfig.io/proxy.conf;
-	}
-
-	include nginxconfig.io/general.conf;
-}
