--- devops/ansible/data/vagrant-nginx/sites-enabled/example.com.conf
@@ -1,1 +1,0 @@
-../sites-available/example.com.conf
