+++ tests/test_row_column.py
@@ -1,7 +1,7 @@
 from flask import url_for
 import pytest
 
-from metabulo.models import CSVFileSchema, db
+from viime.models import CSVFileSchema, db
 
 csv_file_schema = CSVFileSchema()
