+++ tox.ini
@@ -15,7 +15,7 @@ deps = flake8
        flake8-import-order
        flake8-quotes
        pep8-naming
-commands = flake8 {posargs} setup.py tests metabulo
+commands = flake8 {posargs} setup.py tests viime
 
 [testenv:release]
 skip_install = true
@@ -26,7 +26,7 @@ commands = python setup.py sdist
 [flake8]
 format = pylint
 
-application-import-names = metabulo
+application-import-names = viime
 import-order-style = google
 
 max-line-length = 100
