+++ web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "metabulo",
+  "name": "viime",
   "version": "0.1.0",
   "private": true,
   "scripts": {
