+++ web/public/index.html
@@ -6,11 +6,11 @@
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="icon" href="<%= BASE_URL %>favicon.ico">
     <link href="https://fonts.googleapis.com/css?family=Barlow+Condensed" rel="stylesheet">
-    <title>metabulo</title>
+    <title>VIIME</title>
   </head>
   <body>
     <noscript>
-      <strong>We're sorry but metabulo doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
+      <strong>We're sorry but VIIME doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
     </noscript>
     <div id="app"></div>
     <!-- built files will be auto injected -->
